requirements.txt
mediapipe==0.8.7.3
numpy==1.21.5
websockets==8.1
opencv-python
tensorflow

